# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduTrack - Tutor & Coach Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Teachers ──
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List all teachers
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherBody"
      responses:
        "201":
          description: Created teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get a teacher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Not found
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeacherBody"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          description: Not found
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Classes ──
...
  /schedule:
    get:
      operationId: listScheduleSlots
      tags: [schedule]
      summary: List schedule slots
      responses:
        "200":
          description: List of schedule slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleSlot"
    post:
      operationId: createScheduleSlot
      tags: [schedule]
      summary: Create a schedule slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleSlotBody"
      responses:
        "201":
          description: Created schedule slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSlot"
  /schedule/{id}:
    patch:
      operationId: updateScheduleSlot
      tags: [schedule]
      summary: Update a schedule slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleSlotBody"
      responses:
        "200":
          description: Updated schedule slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSlot"
        "404":
          description: Not found
    delete:
      operationId: deleteScheduleSlot
      tags: [schedule]
      summary: Delete a schedule slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        enrolledAt:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - status
        - enrolledAt
        - createdAt
        - updatedAt
    CreateStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        status:
          type: string
          default: active
      required:
        - name
        - email
    UpdateStudentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        status:
          type: string
    Teacher:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        subjects:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - subjects
        - status
        - createdAt
        - updatedAt
    CreateTeacherBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          default: teacher
        subjects:
          type: string
        status:
          type: string
          default: active
      required:
        - name
        - email
        - subjects
    UpdateTeacherBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        subjects:
          type: string
        status:
          type: string
...
    ScheduleSlot:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        className:
          type: string
        teacherName:
          type: ["string", "null"]
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: ["string", "null"]
        color:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - classId
        - className
        - dayOfWeek
        - startTime
        - endTime
        - color
        - createdAt
        - updatedAt
    CreateScheduleSlotBody:
      type: object
      properties:
        classId:
          type: integer
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: ["string", "null"]
      required:
        - classId
        - dayOfWeek
        - startTime
        - endTime
    UpdateScheduleSlotBody:
      type: object
      properties:
        classId:
          type: integer
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: ["string", "null"]