# lib/api-zod/src/generated/types/createTeacherBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduTrack - Tutor & Coach Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeacherBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role?: string;
  subjects: string;
  status?: string;
}