# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduTrack - Tutor & Coach Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStudentBody {
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  grade?: string | null;
  status?: string;
}